+++ .storybook/main.js
@@ -1,7 +1,11 @@
 module.exports = {
   stories: ["../src/**/*.stories.[jt]s", "../src/**/*.stories.[jt]sx", "../src/**/*.stories.mdx"],
   addons: ["@storybook/preset-create-react-app", "@storybook/addon-essentials", "@storybook/addon-a11y"],
-  core: {
-    builder: "webpack5"
-  }
-};
+  framework: {
+    name: "@storybook/react-webpack5",
+    options: {},
+  },
+  docs: {
+    autodocs: true,
+  },
+}