+++ src/Features/Units/UnitSelector/index.spec.tsx
@@ -1,17 +1,23 @@
-import { mount, render } from "enzyme"
 import * as React from "react"
+import { render, screen } from "@testing-library/react"
+import userEvent from "@testing-library/user-event"
 
 import { UnitSystem } from "../types"
 import { UnitSelectorBase } from "./index"
 
 describe("UnitSelector", () => {
-  it("Will display the currently selected unit system", () => {
+  it("Will display the currently selected unit system", async () => {
     const system = UnitSystem.metric
     const switchUnit = jest.fn()
 
-    const wrapper = mount(<UnitSelectorBase system={system} switchUnits={switchUnit} />)
+    render(<UnitSelectorBase system={system} switchUnits={switchUnit} />)
 
     expect(switchUnit).not.toBeCalled()
-    expect(wrapper.text()).toContain("Metric")
+    expect(screen.getAllByRole("button")[0]).toHaveTextContent("Metric")
+    expect(screen.getAllByRole("button")[1]).toHaveTextContent("English")
+
+    const user = userEvent.setup()
+    await user.click(screen.getByText("English"))
+    expect(switchUnit).toHaveBeenCalledWith("English")
   })
 })