+++ src/components/Footer/index.spec.tsx
@@ -1,11 +1,12 @@
 import * as React from "react"
-import * as ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 
 import Footer from "./index"
 
 describe("Footer", () => {
   it("renders without crashing", () => {
     const div = document.createElement("div")
-    ReactDOM.render(<Footer />, div)
+    const root = createRoot(div)
+    root.render(<Footer />)
   })
 })