+++ src/setupTests.ts
@@ -1,8 +1,5 @@
-import Enzyme from "enzyme"
-import Adapter from "enzyme-adapter-react-16"
 import fetchMock, { GlobalWithFetchMock } from "jest-fetch-mock"
-
-Enzyme.configure({ adapter: new Adapter() })
+import "@testing-library/jest-dom"
 
 const customGlobal: GlobalWithFetchMock = global as GlobalWithFetchMock
 customGlobal.fetch = fetchMock